chore: move HBox, VBox classes
