+++ dzgui/views/pages/mods.py
@@ -3,6 +3,7 @@ from pathlib import Path
 from typing import Self, TYPE_CHECKING
 from dzgui.api.steam import find_user_id
 from dzgui.const.enum import NotebookPage, Preferences
+from dzgui.views.components.box import HBox
 from dzgui.views.components.buttons import SteamWorkshopButton
 from dzgui.views.components.scrollable import NoOverlayScrolledWindow
 from dzgui.views.trees.tree_mods import ModTreeView
@@ -36,10 +37,8 @@ class Mods(Gtk.Box):
         uid = find_user_id(steam_path)
 
         pretty_uid = "" if uid is None else uid
-        hbox = Gtk.Box(orientation=Gtk.Orientation.HORIZONTAL, spacing=10)
+        hbox = HBox(spacing=10)
         workshop_button = SteamWorkshopButton()
-        workshop_button.set_margin_top(10)
-        workshop_button.set_margin_bottom(10)
         workshop_button.connect(
             "clicked", lambda _: self.controller.open_user_workshop(pretty_uid)
         )
